--- AstroBaby/View/AstroBaby/View/Base.lproj/Couples.storyboard
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="eI1-Cf-BP1">
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14460.20"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--Couple-->
-        <scene sceneID="Ae2-VK-xBw">
-            <objects>
-                <viewController id="eI1-Cf-BP1" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="bpN-93-6pa">
-                        <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" red="0.023529411760000001" green="0.58431372550000005" blue="0.92549019610000005" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                        <viewLayoutGuide key="safeArea" id="eJQ-N2-DM6"/>
-                    </view>
-                    <tabBarItem key="tabBarItem" title="Couple" id="Q7k-dZ-jJ9"/>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="Phf-Mm-AFR" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="255" y="334"/>
-        </scene>
-    </scenes>
-</document>
